pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.12.15",
    "fastapi>=0.116.1",
    "google-genai>=1.29.0",
    "python-dotenv>=1.1.1",
    "python-telegram-bot>=22.3",
    "sift-stack-py>=0.8.2",
    "uvicorn>=0.35.0",
]
